Refactor : client/config의 socket 관련 상수를 분리
상세 :
- client/constants/socket에서 소켓 관련 상수 관리

 #388
